requirements.txt
django
django-cors-headers
djangorestframework
django-extensions
axios